as_pkgdown {pkgdown}	R Documentation

Generate pkgdown data structure

Description

You will generally not need to use this unless you need a custom site design and you're writing your own equivalent of build_site().

Usage

as_pkgdown(pkg = ".", override = list())

Arguments

pkg
Path to package.

override
An optional named list used to temporarily override values in _pkgdown.yml

[Package pkgdown version 1.1.0 Index]