build_favicons {pkgdown}	R Documentation

Create favicons from package logo

Description

This function auto-detects the location of your package logo (with the name logo.svg (recommended format) or logo.png) and runs it through the https://realfavicongenerator.net API to build a complete set of favicons with different sizes, as needed for modern web usage.

Usage

build_favicons(pkg = ".", overwrite = FALSE)

Arguments

pkg
Path to package.

overwrite
If TRUE, re-create favicons from package logo.

Details

You only need to run the function once. The favicon set will be stored in pkgdown/favicon and copied by init_site() to the relevant location when the website is rebuilt.

Once complete, you should add pkgdown/ to .Rbuildignore to avoid a NOTE during package checking.

[Package pkgdown version 2.0.9 Index]