Build home page

First looks for index.Rmd or README.Rmd, then index.md or README.md. If none are found, falls back to the description field in DESCRIPTION.

build_home(pkg = ".", path = "docs", depth = 0L)

Arguments

pkg

Path to source package.

path

Output path.

depth

Depth of path relative to root of documentation. Used to adjust relative links in the navbar.

YAML config

To tweak the home page, you need a section called home.

The sidebar links are automatically generated by inspecting the URL and BugReports fields of the DESCRIPTION. You can add additional links with a subsection called links, which should contain a list of text + href elements:

home:
  links:
  - text: Link text
    href: http://website.com