Build pkgdown website

build_site() is a convenient wrapper around four functions:

build_articles()

build_home()

build_reference()

build_news()

See the documentation for the each function to learn how to control that aspect of the site.

build_site(pkg = ".", path = "docs", examples = TRUE,
  run_dont_run = FALSE, mathjax = TRUE, preview = interactive(),
  seed = 1014)

Arguments

pkg

Path to source package.

path

Location in which to save website, relative to package path.

examples

Run examples?

run_dont_run

Run examples that are surrounded in \dontrun?

mathjax

Use mathjax to render math symbols?

preview

If TRUE, will preview freshly generated site

seed

Seed used to initialize so that random examples are reproducible.

YAML config

There are four top-level YAML settings that affect the entire site:

title

Site title, used in page title and default navbar.

assets_path

Path to directory of additional assets to be copied into root directory of site. Use this when you want complete control over the visual design of the site. Recommended for advanced users only.

template_path

Path to directory of templates. Override the default (the templates built-in to pkgdown) to completely control the display of your site. Recommended for advanced users only.

template

Additional metadata to be passed on to the template. The default template supports bootswatch to customise the overall theme with a bootswatch theme. You can see a complete list and preview how they look at https://gallery.shinyapps.io/117-shinythemes/.

You can also control the navbar. It uses the same syntax as RMarkdown. The following YAML snippet illustrates some of the most important features.

navbar:
  type: inverse
  left:
    - text: "Home"
      href: index.html
    - text: "Reference"
      href: reference/index.html
    - text: "Articles"
      menu:
        - text: "Heading 1"
        - text: "Article A"
          href: articles/page_a.html
        - text: "Article B"
          href: articles/page_b.html
        - text: "---------"
        - text: "Heading 2"
        - text: "Article C"
          href: articles/page_c.html
        - text: "Article D"
          href: articles/page_d.html
  right:
    - icon: fa-github fa-lg
      href: https://example.com

Use type to choose between "default" and "inverse" themes.

You position elements by placing under either left or right. Components can contain sub-menus with headings (indicated by missing href) and separators. Currently pkgdown only supports fontawesome icons. You can see a full list of options at http://fontawesome.io/icons/.

Examples

## Not run: ------------------------------------
# build_site()
## ---------------------------------------------