check_pkgdown {pkgdown}	R Documentation

Check _pkgdown.yml

Description

Check that your _pkgdown.yml is valid without building the whole site. Currently this:

Checks the reference and article indexes to ensure that pkgdown can read them, and that every documentation topic and vignette/article is included in the index.

Validates any opengraph metadata that you might have supplied

Usage

check_pkgdown(pkg = ".")

Arguments

pkg
Path to package.

[Package pkgdown version 2.0.9 Index]