Clean site

Deletes all files in doc/ (except for CNAME)

clean_site(pkg = ".", path = "docs")

Arguments

pkg
Path to source package. If R working directory is not set to the source directory, then pkg must be a fully qualified path to the source directory (not a relative path).

path
Location in which to save website, relative to package path.