init_site {pkgdown}	R Documentation

Initialise the site

Description

This creates the output directory, creates favicon.ico from package logo, creates a machine readable description of the site, and sets up assets and extra files.

Usage

init_site(pkg = ".")

Arguments

pkg
Path to package.

Custom CSS/JS

If you want to do minor customisation of your pkgdown site, the easiest way is to add pkgdown/extra.css and pkgdown/extra.js. These will be automatically copied to docs/ and inserted into the <HEAD> after the default pkgdown CSS and JSS.

Favicon

If you include you package logo in the standard location of man/figures/logo.png, a favicon will be automatically created for you.

[Package pkgdown version 1.1.0 Index]