pkgdown-package {pkgdown}	R Documentation

pkgdown: Make Static HTML Documentation for a Package

Description

Generate an attractive and useful website from a source package. 'pkgdown' converts your documentation, vignettes, 'README', and more to 'HTML' making it easy to share information about your package online.

Author(s)

Maintainer: Hadley Wickham hadley@posit.co (ORCID)

Authors:

Jay Hesselberth (ORCID)

Maëlle Salmon (ORCID)

Other contributors:

Posit Software, PBC [copyright holder, funder]

See Also

Useful links:

https://pkgdown.r-lib.org

https://github.com/r-lib/pkgdown

Report bugs at https://github.com/r-lib/pkgdown/issues

[Package pkgdown version 2.0.9 Index]