pkgdown_print {pkgdown}	R Documentation

Print object in pkgdown output

Description

This lets package authors control how objects are printed just for pkgdown examples. The default is to call print() apart from htmlwidgets where the object is returned as is (with sizes tweaked).

Usage

pkgdown_print(x, visible = TRUE)

Arguments

x
Object to display

visible
Whether it is visible or not

Value

Either a character vector representing printed output (which will be escaped for HTML as necessary) or literal HTML produced by the htmltools or htmlwidgets packages.

[Package pkgdown version 2.0.9 Index]