template_navbar {pkgdown}	R Documentation

Generate YAML templates

Description

Use these function to generate the default YAML that pkgdown uses for the different parts of _pkgdown.yml. This are useful starting points if you want to customise your site.

Usage

template_navbar(path = ".")

template_reference(path = ".")

template_articles(path = ".")

Arguments

path
Path to package root

[Package pkgdown version 1.1.0 Index]