test-dont {pkgdown}	R Documentation

Test case: don't

Description

Test case: don't

See Also

Other tests: index, test-crayon, test-figures, test-links, test-lists, test-long-lines, test-output-styles, test-params, test-sexpr-title, test-verbatim

Examples

## Not run: 
  abort("This is an error!")

## End(Not run)

# Inline \donttest is silently ommitted
message("Hi!")

# Block \donttest indicated with comments

# This is a comment
1 + 3

# And works even when not at the top level
if (TRUE) {
  
  1 + 2
  
}

answer <- 1

answer # should be 42

# To hide the \dontshow part, for conditional examples

answer <- 43

answer # should be still 42

# But this one runs, and the condition is hidden

answer <- 43

answer

[Package pkgdown version 2.0.9 Index]