test-verbatim {pkgdown}	R Documentation

Test case: preformatted blocks & syntax highlighting

Description

Manual test cases for various ways of embedding code in sections. All code blocks should have copy and paste button.

Should be highlighted

Valid R code in \preformatted{}:

mean(a + 1)

R code in R block:

mean(a + 1)

R code in r block:

mean(a + 1)

Yaml

yaml: [a, 1]

Shouldn't be highlighted

Non-R code in \preformatted{}

yaml: [a, b, c]

See Also

Other tests: index, test-crayon, test-dont, test-figures, test-links, test-lists, test-long-lines, test-output-styles, test-params, test-sexpr-title

[Package pkgdown version 2.0.9 Index]