Build home section

This function is responsible for building .md (or .Rmd) files typically found in the root directory of the package. It will generate the home page from index.Rmd or README.Rmd, then index.md or README.md. If none of these are found, it falls back to the description field in DESCRIPTION. It also builds any files found in .github/.

build_home(pkg = ".", override = list(), preview = NA,
  quiet = TRUE, devel = TRUE)

Arguments

pkg
Path to package.

override
An optional named list used to temporarily override values in _pkgdown.yml

preview
If TRUE, or is.na(preview) && interactive(), will preview freshly generated section in browser.

quiet
Set to FALSE to display output of knitr and pandoc. This is useful when debugging.

devel
If TRUE, assumes you are doing active development, and re-knits all .Rmd files. If FALSE, assumes Rmarkdown files have already rendered to .md, falling back to rendering as plain markdown if they have not.

Images and figures

If you want to include images in your README.md, they must be stored somewhere in the package so that they can be displayed on the CRAN website. The best place to put them is man/figures. If you are generating figures with R Markdown, make sure you set up fig.path as followed:

```{r, include = FALSE}
knitr::opts_chunk$set(
  fig.path = "man/figures/"
)
```

Package logo

If you have a package logo, you can include it at the top of your README in a level-one heading:

# pkgdown <img src="man/figures/logo.png" align="right" />

init_site() will also automatically create a favicon set from your package logo.

YAML config - home

To tweak the home page, you need a section called home.

The sidebar links are automatically generated by inspecting the URL and BugReports fields of the DESCRIPTION. You can add additional links with a subsection called links, which should contain a list of text + href elements:

home:
  links:
  - text: Link text
    href: http://website.com

You can remove the first heading with:

home:
  strip_header: true

YAML config - home title and description

The homepage title that will be displayed in the browser tab, and the title of homepage in Twitter/Open Graph metadata ("og:title"), is by default the package Title from DESCRIPTION.

The description of the homepage in Twitter/Open Graph metadata ("og:description") is by default the package Description from DESCRIPTION, with words quoted with '' unquoted.

You can override these values by adding a home section in the pkgdown config file, and add title and/or description, with words quoted by '' unquoted.

home:
  title: This is my package title
  description: This is a free description of my package

YAML config - authors

The "developers" list is populated by the maintainer ("cre"), authors ("aut"), and funder ("fnd") from the DESCRIPTION. You can modify their display on the home page by adding a subsection for authors. The author name should match the value in DESCRIPTION:

authors:
  name:
    href: "http://name-website.com"
    html: "<img src='name-picture.png' height=24>"

ORCID

ORCID identification numbers are added to author info if they are defined in the DESCRIPTION:

person("Jay", "Hesselberth", role = "aut",
  comment = c(ORCID = "0000-0002-6299-179X")
),

Badges

Status badges are displayed in the sidebar under the section "Dev status". This section is automatically populated if there is an identifiable badges paragraph in the homepage source (index.Rmd, index.md, README.Rmd, README.md):

pkgdown first looks for a paragraph starting with <!-- badges: start --> and ending with <!-- badges: end --> as created by usethis::use_readme_md() or usethis::use_readme_rmd();

failing that, pkgdown looks at the first paragraph, and uses it if it only contains images.