Build news section

Your NEWS.md is parsed in to sections based on your use of headings.

build_news(pkg = ".", override = list(), preview = NA)

Arguments

pkg
Path to package.

override
An optional named list used to temporarily override values in _pkgdown.yml

preview
If TRUE, or is.na(preview) && interactive(), will preview freshly generated section in browser.

Details

The NEWS.md file should be formatted with level one headings (#) containing the package name and version number, level two headings (##) with topic headings and lists of news bullets. Commonly used level two headings include 'Major changes', 'Bug fixes', or 'Minor changes'.

# pkgdown 0.1.0.9000

## Major changes

- Fresh approach based on the staticdocs package. Site configuration now based
  on YAML files.

If the package is available on CRAN, release dates will be added to versions in level-one headings, and "Unreleased" will be added versions that are not on CRAN.

Issues and contributors mentioned in news items are automatically linked to github if a URL entry linking to github.com is provided in the package DESCRIPTION.

## Major changes

- Lots of bug fixes (@hadley, #100)

YAML config

To automatically link to release announcements, include a releases section.

news:
 releases:
 - text: "usethis 1.3.0"
   href: https://www.tidyverse.org/articles/2018/02/usethis-1-3-0/
 - text: "usethis 1.0.0 (and 1.1.0)"
   href: https://www.tidyverse.org/articles/2017/11/usethis-1.0.0/

Control whether news is present on one page or multiple pages with the one_page field. The default is true.

news:
- one_page: false

See also