Build reference section

By default, pkgdown will generate an index that lists all functions in alphabetical order. To override this, provide a reference section in your _pkgdown.yml as described below.

build_reference(
  pkg = ".",
  lazy = TRUE,
  examples = TRUE,
  run_dont_run = FALSE,
  seed = 1014,
  override = list(),
  preview = NA,
  devel = TRUE,
  document = "DEPRECATED",
  topics = NULL
)

build_reference_index(pkg = ".")

Arguments

pkg
Path to package.

lazy
If TRUE, only rebuild pages where the .Rd is more recent than the .html. This makes it much easier to rapidly prototype. It is set to FALSE by build_site().

examples
Run examples?

run_dont_run
Run examples that are surrounded in \dontrun?

seed
Seed used to initialize so that random examples are reproducible.

override
An optional named list used to temporarily override values in _pkgdown.yml

preview
If TRUE, or is.na(preview) && interactive(), will preview freshly generated section in browser.

devel
Determines how code is loaded in order to run examples. If TRUE (the default), assumes you are in a live development environment, and loads source package with pkgload::load_all(). If FALSE, uses the installed version of the package.

document
Deprecated Use devel instead.

topics
Build only specified topics. If supplied, sets lazy`` and previewtoFALSE`.

YAML config

To tweak the index page, add a section called reference to _pkgdown.yml which provides a list of sections containing, a title, list of contents, and an optional description.

For example, the following code breaks up the functions in pkgdown into two groups:

reference:
- title: Render components
  desc:  Build each component of the site.
  contents:
  - starts_with("build_")
  - init_site
- title: Templates
  contents:
  - render_page

contents can contain:

Individual function/topic names.

Weirdly named functions with doubled quoting, once for YAML and once for R, e.g. "`+.gg`".

starts_with("prefix") to select all functions with common prefix.

ends_with("suffix") to select all functions with common suffix.

matches("regexp") for more complex regular expressions.

has_keyword("x") to select all topics with keyword "x"; has_keyword("datasets") selects all data documentation.

has_concept("blah") to select all topics with concept "blah".

lacks_concepts(c("concept1", "concept2")) to select all topics without those concepts. This is useful to capture topics not otherwise captured by has_concepts().

All functions (except for has_keywords()) automatically exclude internal topics (i.e. those with \keyword{internal}). You can choose to include with (e.g.) starts_with("build_", internal = TRUE).

Use a leading - to remove topics from a section, e.g. -topic_name, -starts_with("foo").

You can provide long descriptions for groups of functions using the YAML > notation:

desc: >
  This is a very long and overly flowery description of a
  single simple function.

pkgdown will check that all non-internal topics are included on this page, and will generate a warning if you have missed any.

Figures

You can control the default rendering of figures by specifying the figures field in _pkgdown.yml. The default settings are equivalent to:

figures:
  dev: grDevices::png
  dpi: 96
  dev.args: []
  fig.ext: png
  fig.width: 7.2916667
  fig.height: ~
  fig.retina: 2
  fig.asp: 1.618

Icons

You can optionally supply an icon for each help topic. To do so, you'll need a top-level icons directory. This should contain .png files that are either 30x30 (for regular display) or 60x60 (if you want retina display). Icons are matched to topics by aliases.