Build and deploy a site locally

Assumes that you're in a git clone of the project, and the package is already installed. Use usethis::use_pkgdown_github_pages() to automate this process using GitHub actions.

Usage

deploy_to_branch(
  pkg = ".",
  commit_message = construct_commit_message(pkg),
  clean = TRUE,
  branch = "gh-pages",
  remote = "origin",
  github_pages = (branch == "gh-pages"),
  ...
)

Arguments

pkg

Path to package.

commit_message

The commit message to be used for the commit.

clean

Clean all files from old site.

branch

The git branch to deploy to

remote

The git remote to deploy to

github_pages

Is this a GitHub pages deploy. If TRUE, adds a CNAME file for custom domain name support, and a .nojekyll file to suppress jekyll rendering.

...

Additional arguments passed to build_site().