init_site {pkgdown}	R Documentation

Initialise site infrastructure

Description

init_site():

creates the output directory (docs/),

generates a machine readable description of the site, used for autolinking,

copies CSS/JS assets and extra files, and

runs build_favicons(), if needed.

See vignette("customise") for the various ways you can customise the display of your site.

Usage

init_site(pkg = ".")

Arguments

pkg
Path to package.

Build-ignored files

We recommend using usethis::use_pkgdown() to build-ignore docs/ and _pkgdown.yml. If use another directory, or create the site manually, you'll need to add them to .Rbuildignore yourself. A NOTE about an unexpected file during R CMD CHECK is an indication you have not correctly ignored these files.

[Package pkgdown version 2.0.9 Index]